requirements.txt
python-telegram-bot==13.11
boto3==1.21.16